client concrete function in separate module
