+++ src/qibo_client/__init__.py
@@ -1,2 +1,2 @@
 """The `qibo_tii_provider` package"""
-from qibo_client.qibo_client import TII
+from .tii import TII
